Address CodeRabbit review; add ASCII logo; latest Go/deps; perf notes
CodeRabbit review fixes:
- config.Addr: use net.SplitHostPort so bare IPv6 tailnet hosts get :22.
- render.Headless / stdio: check eng.Write and term.Restore errors.
- sshx.hostkey: reload known_hosts after TOFU append (no stale snapshot /
  duplicate entries); serialize the callback.
- sshx.Dial: drain the handshake goroutine on ctx cancel so a live client
  can't leak.
- terminal/utf8: a byte that breaks a multi-byte sequence is now reprocessed
  (ESC/ASCII/control no longer dropped); regression test added.
- update.Apply: Windows-safe replace (rename-aside + rollback) instead of
  os.Rename over the running binary.
- update_test: exercise the real checksum-mismatch branch (wrong hash, not a
  404).
- transport/tsnet: route interactive auth URL to UserLogf, not Logf.
- cmd/wisp: in-app update install honours the run ctx (cancellable).
- testutil/sshserver: uint64 bounds check in parsePtyReq.
- CI/release: pin actions to commit hashes, persist-credentials: false,
  least-privilege permissions, -race for the GUI tests, run ebiten tests on
  macOS, gh release create instead of a third-party action, ad-hoc codesign
  of the macOS artifacts. README MD040 fences; richer .gitignore.

Also picked up the unique cases from CodeRabbit's test PRs (#2/#3): controller
dial-failure/Close/Stderr, headless non-EOF error, snapshot independence and
resize edge cases.

New per user request:
- Latest Go (1.26.4) aligned across go.mod/CI/docs; updated direct deps
  (x/image 0.43). gvisor intentionally left at tailscale's pin.
- ASCII startup logo (internal/banner): printed to stderr (constant string,
  zero startup cost) and shown as a GUI splash until the first remote byte.
- Performance: added VT-engine throughput/snapshot benchmarks and an honest
  README "Performance" section — pure-Go engine is fine for interactive use;
  Ghostty-class speed comes from the -tags libghostty engine + -tags ebiten
  GPU rendering.

+++ internal/testutil/sshserver/sshserver.go
@@ -235,7 +235,9 @@ func parsePtyReq(b []byte) (term string, cols, rows uint32, ok bool) {
 		return "", 0, 0, false
 	}
 	n := beUint32(b)
-	if int(n)+4+8 > len(b) {
+	// Validate with a wide unsigned type so a peer-controlled length cannot
+	// overflow int (and bypass the bounds check) on 32-bit platforms.
+	if uint64(n)+4+8 > uint64(len(b)) {
 		return "", 0, 0, false
 	}
 	term = string(b[4 : 4+n])
